+++ src/server.js
@@ -13,11 +13,13 @@ const clientRoutes = require('./routes/client');
 const adminRoutes = require('./routes/admin');
 const appearanceRoutes = require('./routes/appearance');
 const appearance = require('./services/appearance');
+const automations = require('./services/automations');
 const consoleWs = require('./ws/console');
 const sftp = require('./sftp/sftpServer');
 const { securityHeaders } = require('./middleware');
 
 db.load();
+automations.init(); // start watching consoles for servers that have rules
 
 const app = express();
 app.disable('x-powered-by');